asset: Imported Seed, paper, and garbage hint animation
Imported and applied animation spritesheets of seed, paper and garbage hints and assigned them to the objects SeedAnimation, GarbageAnimation, and PaperAnimation in the scene ImageTracking. Also changed the sprite sheet limit from 2048 to 4096 to prevent blurry animation. Resized container for animation to better scale within scene. (Note: container size might have to be iterated on later)