feat: Full compass transition animation
CompassTransition in ImageTracking is at a lowered state at all times on enable. Change the value of sceneprogress in CompassTransition.controller to 1-5 for the correct animation for each scene transition.
The script resetSceneProgressValue resets sceneprogress at the end of the animation. So triggering a reveal only requires changing that value